FIXED-If defense is empty, there shouldn't be a battle report. I got one and it was all messed up because there was no defense. This might happen if town center is level 1 and there are no units. FIXED-Is tower joining batle properly? FIXED-Lumbermill needs description about which lvl forest it can gather wood from. FIXED-Build time decreased based on main building level FIXED-Add icons to properties FIXED-Remove action count from all tooltips (the 0/1 at the bottom left) FIXED-Can we sort the name of the troops in the battle report? so the tabs don't switch places. Checked with Anthony: They stack but have a upper limit. -How do the armor/weapon research technologies stack? FIXED-Do any technologies decrease training time or building time? Client needs those logics too. FIXED-Organize formulas in server FIXED-Technologies will not get re-added if structure is re-upgraded after being downgraded and Init tries to add it again. FIXED-Add OUT OF STAMINA to battle reasons for website FIXED-Adjusted how much unit speed helps against bow units.

+++ ConsoleSimulator/Simulation.cs
@@ -153,9 +153,9 @@ namespace ConsoleSimulator {
                 obj.HitDealt, 
                 obj.HitRecv,
                 obj.MaxDmgDealt,
-                obj.MinDmgDealt==int.MaxValue?0:obj.MinDmgDealt, 
+                obj.MinDmgDealt==ushort.MaxValue?0:obj.MinDmgDealt, 
                 obj.MaxDmgRecv,
-                obj.MinDmgRecv==int.MaxValue?0:obj.MinDmgRecv,
+                obj.MinDmgRecv==ushort.MaxValue?0:obj.MinDmgRecv,
                 obj.Count);
         }
